Moved the new fluids.py functionality to case.
Updated test_case.py to reflect this
